feat: Implement useForm hook for handling registration and login submissions
